requirements.txt
beautifulsoup4>=4.12.2
lxml>=4.9.3
Pillow>=10.0.0
requests>=2.31.0
streamlit>=1.30.0
pyperclip>=1.8.2
selenium>=4.15.0
webdriver-manager>=4.0.0
